Local_Ebot integration begun in Fetchutil

+++ Fetchutil.py
@@ -7,30 +7,52 @@ import errno
 import time
 
 def seqfetch(acc):
-    "prints the sequence in fasta format to file"
+  "prints the sequence in fasta format to file"
+  if os.path.exists('Proteomes/orgmap_all'):
+    dict_all={}
+    with open('Proteomes/orgmap_all') as infile:
+      for line in infile:
+        spl=line.split()
+        dict_all.update({spl[0]:spl[1:]})
+    taxid=dict_all[acc][0]
+    taxall={}
+    with open('Proteomes/all_tax') as infile:
+      for line in infile:
+        spl=line.split('\t')
+        taxall.update({spl[0]:spl[1]})
+      binom=taxall[taxid]
+      bspl=binom.split()
+      four=bspl[0][0]+bspl[1][:3]
+      four.lower()
+      ##DB params
+      db='Proteomes/'+four
+      out='Orthos/'+acc+'.fas'
+      subprocess.call(['blastdbcmd','-db',db,'-out',out,'-entry',acc])
+      return out
+  else:
     hand=Entrez.efetch(db='protein',id=str(acc),rettype='fasta')
     string=''
     hand.readline()
     org=orgfetch(acc)
     string+='>'+org[0]+': '+org[1]+' '+acc+'\n'
     while hand:
-        inpout=hand.readline()
-        string+=inpout
-        if inpout=='':
-            if not os.path.exists('Orthos'):
-              subprocess.call(['mkdir','Orthos'])
-            open('Orthos/'+acc+'.fasta','w').write(string)
-            return 'Orthos/'+acc+'.fasta'
+      inpout=hand.readline()
+      string+=inpout
+      if inpout=='':
+        if not os.path.exists('Orthos'):
+          subprocess.call(['mkdir','Orthos'])
+          open('Orthos/'+acc+'.fasta','w').write(string)
+          return 'Orthos/'+acc+'.fasta'
 
 def addseq(oldseq, newseq):
-    "Transfers the sequence from newseq to oldseq"
-    old=open(oldseq,'a')
-    new=open(newseq)
-    while new:
-        add=new.readline()
-        old.write(add)
-        if add=='':
-            return
+  "Transfers the sequence from newseq to oldseq"
+  old=open(oldseq,'a')
+  new=open(newseq)
+  while new:
+    add=new.readline()
+    old.write(add)
+    if add=='':
+      return
 def namefetch(acc):
   "Returns the description of acc"
   hand=Entrez.esummary(db='protein',id=acc)
@@ -65,6 +87,28 @@ def namefetch(acc):
 def orgfetch(acc):
   "Fetches source organism for acc"
   ret=[]
+  if os.path.exists('Proteomes/orgmap_all'):
+    dict_all={}
+    with open('Proteomes/orgmap_all') as infile:
+      for line in infile:
+        spl=line.split()
+        dict_all.update({spl[0]:spl[1:]})
+    taxid=dict_all[acc][0]
+    taxall={}
+    with open('Proteomes/all_tax') as infile:
+      for line in infile:
+        spl=line.split('\t')
+        taxall.update({spl[0]:spl[1]})
+      binom=taxall[taxid]
+      ret.append(binom)
+      bspl=binom.split()
+      four=bspl[0][0]+bspl[1][:3]
+      four.lower()
+  with Entrez.efetch(db='taxonomy',id=ids) as hand:
+    for line in hand:
+      with open(four+'_tax_fetch','w') as taxf:
+        taxf.write(line)
+
   try:
     hand=Entrez.esummary(db="protein", id=acc)
   except SocketError:
